Forecasting simulation

Creating different forecast paths for forecast objects (when applicable), by utilizing the underline model distribution with the simulate function

forecast_sim(model, h, n, sim_color = "blue", opacity = 0.05,
  plot = TRUE)

Arguments

model

A forecasting model supporting Arima, auto.arima, ets, and nnetar models from the **forecast** package

h

An integer, defines the forecast horizon

n

An integer, set the number of iterations of the simulation

sim_color

Set the color of the simulation paths lines

opacity

Set the opacity level of the simulation path lines

plot

Logical, if TRUE will desplay the output plot

Value

The baseline series, the simulated values and a plot

Examples

 if (FALSE) {
library(forecast)
data(USgas)

# Create a model
fit <- auto.arima(USgas)

# Simulate 100 possible forecast path, with horizon of 60 months
forecast_sim(model = fit, h = 60, n = 100)
}