Plotting Forecast Object

Visualization functions for forecast package forecasting objects

plot_forecast(forecast_obj, title = NULL, Xtitle = NULL,
  Ytitle = NULL, color = NULL, width = 2)

Arguments

forecast_obj

A forecast object from the forecast, forecastHybrid, or bsts packages

title

A character, a plot title, optional

Xtitle

Set the X axis title, default set to NULL

Ytitle

Set the Y axis title, default set to NULL

color

A character, the plot, support both name and expression

width

An Integer, define the plot width, default is set to 2

Examples

data(USgas)
library(forecast)
fit <- ets(USgas)
fc<- forecast(fit, h = 60)
plot_forecast(fc)