Train, Test, Evaluate, and Forecast Multiple Time Series Forecasting Models

Method for train test and compare multiple time series models using either one partition (i.e., sample out) or multipe partitions (backtesting)

train_model(input, methods, train_method, horizon, error = "MAPE",
  xreg = NULL)

Arguments

input
A univariate time series object (ts class)

methods
A list, defines the models to use for training and forecasting the series. The list must include a sub list with the model type, and the model's arguments (when applicable) and notes about the model. The sub-list name will be used as the model ID. Possible models:

arima - model from the stats package

auto.arima - model from the forecast package

ets - model from the forecast package

HoltWinters - model from the stats package

nnetar - model from the forecast package

tslm - model from the forecast package (note that the 'tslm' model must have the formula argument in the 'method_arg' argument)

train_method
A list, defines the backtesting parameters:

partitions - an integer, set the number of training and testing partitions to be used in the backtesting process, where when partition is set to 1 it is a simple holdout training approach

space - an integer, defines the length of the backtesting window expansion

sample.in - an integer, optional, defines the length of the training partitions, and therefore the backtesting window structure. By default, it set to NULL and therefore, the backtesting using expending window. Otherwise, when the sample.in defined, the window structure is sliding

sample.in - an integer, optional, defines the length of the training partitions, and therefore the type of the backtesting window. By default, is set to NULL, which implay that the backtesting is using an expending window. Otherwise, when defining the size of the training partition, th defines the train approach, either using a single testing partition (sample out) or use multiple testing partitions (backtesting). The list should include the training method argument, (please see 'details' for the structure of the argument)

horizon
An integer, defines the forecast horizon

error
A character, defines the error metrics to be used to sort the models leaderboard. Possible metric - "MAPE" or "RMSE"

xreg
Optional, a list with two vectors (e.g., data.frame or matrix) of external regressors, one vector corresponding to the input series and second to the forecast itself (e.g., must have the same length as the input and forecast horizon, respectively)

Examples

# Defining the models and their arguments
methods <- list(ets1 = list(method = "ets",
                            method_arg = list(opt.crit = "lik"),
                            notes = "ETS model with opt.crit = lik"),
                ets2 = list(method = "ets",
                            method_arg = list(opt.crit = "amse"),
                            notes = "ETS model with opt.crit = amse"),
                arima1 = list(method = "arima",
                              method_arg = list(order = c(2,1,1),
                                                seasonal = list(order = c(1,0,1))),
                              notes = "SARIMA(2,1,1)(1,0,1)"),
                arima2 = list(method = "arima",
                              method_arg = list(order = c(2,1,2),
                                                seasonal = list(order = c(1,1,1))),
                              notes = "SARIMA(2,1,2)(1,1,1)"),
                auto_arima = list(method = "auto.arima",
                                  method_arg = NULL,
                                  notes = "auto.arima model"),
                hw = list(method = "HoltWinters",
                          method_arg = NULL,
                          notes = "HoltWinters Model"),
                tslm = list(method = "tslm",
                            method_arg = list(formula = input ~ trend + season),
                            notes = "tslm model with trend and seasonal components"))

# Training the models with backtesting
md <- train_model(input = USgas,
                  methods = methods,
                  train_method = list(partitions = 6,
                                      sample.out = 12,
                                      space = 3),
                  horizon = 12,
                  error = "MAPE")

#> Warning: NaNs produced

#> Error in (function (x, order = c(0L, 0L, 0L), seasonal = list(order = c(0L,     0L, 0L), period = NA), xreg = NULL, include.mean = TRUE,     transform.pars = TRUE, fixed = NULL, init = NULL, method = c("CSS-ML",         "ML", "CSS"), n.cond, SSinit = c("Gardner1980", "Rossignol2011"),     optim.method = "BFGS", optim.control = list(), kappa = 1e+06) {    "%+%" <- function(a, b) .Call(C_TSconv, a, b)    SSinit <- match.arg(SSinit)    SS.G <- SSinit == "Gardner1980"    upARIMA <- function(mod, phi, theta) {        p <- length(phi)        q <- length(theta)        mod$phi <- phi        mod$theta <- theta        r <- max(p, q + 1L)        if (p > 0)             mod$T[1L:p, 1L] <- phi        if (r > 1L)             mod$Pn[1L:r, 1L:r] <- if (SS.G)                 .Call(C_getQ0, phi, theta)            else .Call(C_getQ0bis, phi, theta, tol = 0)        else mod$Pn[1L, 1L] <- if (p > 0)             1/(1 - phi^2)        else 1        mod$a[] <- 0        mod    }    arimaSS <- function(y, mod) {        .Call(C_ARIMA_Like, y, mod, 0L, TRUE)    }    armafn <- function(p, trans) {        par <- coef        par[mask] <- p        trarma <- .Call(C_ARIMA_transPars, par, arma, trans)        if (is.null(Z <- tryCatch(upARIMA(mod, trarma[[1L]],             trarma[[2L]]), error = function(e) NULL)))             return(.Machine$double.xmax)        if (ncxreg > 0)             x <- x - xreg %*% par[narma + (1L:ncxreg)]        res <- .Call(C_ARIMA_Like, x, Z, 0L, FALSE)        s2 <- res[1L]/res[3L]        0.5 * (log(s2) + res[2L]/res[3L])    }    armaCSS <- function(p) {        par <- as.double(fixed)        par[mask] <- p        trarma <- .Call(C_ARIMA_transPars, par, arma, FALSE)        if (ncxreg > 0)             x <- x - xreg %*% par[narma + (1L:ncxreg)]        res <- .Call(C_ARIMA_CSS, x, arma, trarma[[1L]], trarma[[2L]],             as.integer(ncond), FALSE)        0.5 * log(res)    }    arCheck <- function(ar) {        p <- max(which(c(1, -ar) != 0)) - 1        if (!p)             return(TRUE)        all(Mod(polyroot(c(1, -ar[1L:p]))) > 1)    }    maInvert <- function(ma) {        q <- length(ma)        q0 <- max(which(c(1, ma) != 0)) - 1L        if (!q0)             return(ma)        roots <- polyroot(c(1, ma[1L:q0]))        ind <- Mod(roots) < 1        if (all(!ind))             return(ma)        if (q0 == 1)             return(c(1/ma[1L], rep.int(0, q - q0)))        roots[ind] <- 1/roots[ind]        x <- 1        for (r in roots) x <- c(x, 0) - c(0, x)/r        c(Re(x[-1L]), rep.int(0, q - q0))    }    series <- deparse(substitute(x))    if (NCOL(x) > 1L)         stop("only implemented for univariate time series")    method <- match.arg(method)    x <- as.ts(x)    if (!is.numeric(x))         stop("'x' must be numeric")    storage.mode(x) <- "double"    dim(x) <- NULL    n <- length(x)    if (!missing(order))         if (!is.numeric(order) || length(order) != 3L || any(order <             0))             stop("'order' must be a non-negative numeric vector of length 3")    if (!missing(seasonal))         if (is.list(seasonal)) {            if (is.null(seasonal$order))                 stop("'seasonal' must be a list with component 'order'")            if (!is.numeric(seasonal$order) || length(seasonal$order) !=                 3L || any(seasonal$order < 0L))                 stop("'seasonal$order' must be a non-negative numeric vector of length 3")        }        else if (is.numeric(order)) {            if (length(order) == 3L)                 seasonal <- list(order = seasonal)            else ("'seasonal' is of the wrong length")        }        else stop("'seasonal' must be a list with component 'order'")    if (is.null(seasonal$period) || is.na(seasonal$period) ||         seasonal$period == 0)         seasonal$period <- frequency(x)    arma <- as.integer(c(order[-2L], seasonal$order[-2L], seasonal$period,         order[2L], seasonal$order[2L]))    narma <- sum(arma[1L:4L])    xtsp <- tsp(x)    tsp(x) <- NULL    Delta <- 1    for (i in seq_len(order[2L])) Delta <- Delta %+% c(1, -1)    for (i in seq_len(seasonal$order[2L])) Delta <- Delta %+%         c(1, rep.int(0, seasonal$period - 1), -1)    Delta <- -Delta[-1L]    nd <- order[2L] + seasonal$order[2L]    n.used <- sum(!is.na(x)) - length(Delta)    if (is.null(xreg)) {        ncxreg <- 0L    }    else {        nmxreg <- deparse(substitute(xreg))        if (NROW(xreg) != n)             stop("lengths of 'x' and 'xreg' do not match")        ncxreg <- NCOL(xreg)        xreg <- as.matrix(xreg)        storage.mode(xreg) <- "double"    }    class(xreg) <- NULL    if (ncxreg > 0L && is.null(colnames(xreg)))         colnames(xreg) <- if (ncxreg == 1L)             nmxreg        else paste0(nmxreg, 1L:ncxreg)    if (include.mean && (nd == 0L)) {        xreg <- cbind(intercept = rep(1, n), xreg = xreg)        ncxreg <- ncxreg + 1L    }    if (method == "CSS-ML") {        anyna <- anyNA(x)        if (ncxreg)             anyna <- anyna || anyNA(xreg)        if (anyna)             method <- "ML"    }    if (method == "CSS" || method == "CSS-ML") {        ncond <- order[2L] + seasonal$order[2L] * seasonal$period        ncond1 <- order[1L] + seasonal$period * seasonal$order[1L]        ncond <- ncond + if (!missing(n.cond))             max(n.cond, ncond1)        else ncond1    }    else ncond <- 0    if (is.null(fixed))         fixed <- rep(NA_real_, narma + ncxreg)    else if (length(fixed) != narma + ncxreg)         stop("wrong length for 'fixed'")    mask <- is.na(fixed)    no.optim <- !any(mask)    if (no.optim)         transform.pars <- FALSE    if (transform.pars) {        ind <- arma[1L] + arma[2L] + seq_len(arma[3L])        if (any(!mask[seq_len(arma[1L])]) || any(!mask[ind])) {            warning("some AR parameters were fixed: setting transform.pars = FALSE")            transform.pars <- FALSE        }    }    init0 <- rep.int(0, narma)    parscale <- rep(1, narma)    if (ncxreg) {        cn <- colnames(xreg)        orig.xreg <- (ncxreg == 1L) || any(!mask[narma + 1L:ncxreg])        if (!orig.xreg) {            S <- svd(na.omit(xreg))            xreg <- xreg %*% S$v        }        dx <- x        dxreg <- xreg        if (order[2L] > 0L) {            dx <- diff(dx, 1L, order[2L])            dxreg <- diff(dxreg, 1L, order[2L])        }        if (seasonal$period > 1L & seasonal$order[2L] > 0) {            dx <- diff(dx, seasonal$period, seasonal$order[2L])            dxreg <- diff(dxreg, seasonal$period, seasonal$order[2L])        }        fit <- if (length(dx) > ncol(dxreg))             lm(dx ~ dxreg - 1, na.action = na.omit)        else list(rank = 0L)        if (fit$rank == 0L) {            fit <- lm(x ~ xreg - 1, na.action = na.omit)        }        isna <- is.na(x) | apply(xreg, 1L, anyNA)        n.used <- sum(!isna) - length(Delta)        init0 <- c(init0, coef(fit))        ses <- summary(fit)$coefficients[, 2L]        parscale <- c(parscale, 10 * ses)    }    if (n.used <= 0)         stop("too few non-missing observations")    if (!is.null(init)) {        if (length(init) != length(init0))             stop("'init' is of the wrong length")        if (any(ind <- is.na(init)))             init[ind] <- init0[ind]        if (method == "ML") {            if (arma[1L] > 0)                 if (!arCheck(init[1L:arma[1L]]))                   stop("non-stationary AR part")            if (arma[3L] > 0)                 if (!arCheck(init[sum(arma[1L:2L]) + 1L:arma[3L]]))                   stop("non-stationary seasonal AR part")            if (transform.pars)                 init <- .Call(C_ARIMA_Invtrans, as.double(init),                   arma)        }    }    else init <- init0    coef <- as.double(fixed)    if (!("parscale" %in% names(optim.control)))         optim.control$parscale <- parscale[mask]    if (method == "CSS") {        res <- if (no.optim)             list(convergence = 0L, par = numeric(), value = armaCSS(numeric()))        else optim(init[mask], armaCSS, method = optim.method,             hessian = TRUE, control = optim.control)        if (res$convergence > 0)             warning(gettextf("possible convergence problem: optim gave code = %d",                 res$convergence), domain = NA)        coef[mask] <- res$par        trarma <- .Call(C_ARIMA_transPars, coef, arma, FALSE)        mod <- makeARIMA(trarma[[1L]], trarma[[2L]], Delta, kappa,             SSinit)        if (ncxreg > 0)             x <- x - xreg %*% coef[narma + (1L:ncxreg)]        arimaSS(x, mod)        val <- .Call(C_ARIMA_CSS, x, arma, trarma[[1L]], trarma[[2L]],             as.integer(ncond), TRUE)        sigma2 <- val[[1L]]        var <- if (no.optim)             numeric()        else solve(res$hessian * n.used)    }    else {        if (method == "CSS-ML") {            res <- if (no.optim)                 list(convergence = 0L, par = numeric(), value = armaCSS(numeric()))            else optim(init[mask], armaCSS, method = optim.method,                 hessian = FALSE, control = optim.control)            if (res$convergence == 0)                 init[mask] <- res$par            if (arma[1L] > 0)                 if (!arCheck(init[1L:arma[1L]]))                   stop("non-stationary AR part from CSS")            if (arma[3L] > 0)                 if (!arCheck(init[sum(arma[1L:2L]) + 1L:arma[3L]]))                   stop("non-stationary seasonal AR part from CSS")            ncond <- 0L        }        if (transform.pars) {            init <- .Call(C_ARIMA_Invtrans, init, arma)            if (arma[2L] > 0) {                ind <- arma[1L] + 1L:arma[2L]                init[ind] <- maInvert(init[ind])            }            if (arma[4L] > 0) {                ind <- sum(arma[1L:3L]) + 1L:arma[4L]                init[ind] <- maInvert(init[ind])            }        }        trarma <- .Call(C_ARIMA_transPars, init, arma, transform.pars)        mod <- makeARIMA(trarma[[1L]], trarma[[2L]], Delta, kappa,             SSinit)        res <- if (no.optim)             list(convergence = 0, par = numeric(), value = armafn(numeric(),                 as.logical(transform.pars)))        else optim(init[mask], armafn, method = optim.method,             hessian = TRUE, control = optim.control, trans = as.logical(transform.pars))        if (res$convergence > 0)             warning(gettextf("possible convergence problem: optim gave code = %d",                 res$convergence), domain = NA)        coef[mask] <- res$par        if (transform.pars) {            if (arma[2L] > 0L) {                ind <- arma[1L] + 1L:arma[2L]                if (all(mask[ind]))                   coef[ind] <- maInvert(coef[ind])            }            if (arma[4L] > 0L) {                ind <- sum(arma[1L:3L]) + 1L:arma[4L]                if (all(mask[ind]))                   coef[ind] <- maInvert(coef[ind])            }            if (any(coef[mask] != res$par)) {                oldcode <- res$convergence                res <- optim(coef[mask], armafn, method = optim.method,                   hessian = TRUE, control = list(maxit = 0L,                     parscale = optim.control$parscale), trans = TRUE)                res$convergence <- oldcode                coef[mask] <- res$par            }            A <- .Call(C_ARIMA_Gradtrans, as.double(coef), arma)            A <- A[mask, mask]            var <- crossprod(A, solve(res$hessian * n.used, A))            coef <- .Call(C_ARIMA_undoPars, coef, arma)        }        else var <- if (no.optim)             numeric()        else solve(res$hessian * n.used)        trarma <- .Call(C_ARIMA_transPars, coef, arma, FALSE)        mod <- makeARIMA(trarma[[1L]], trarma[[2L]], Delta, kappa,             SSinit)        val <- if (ncxreg > 0L)             arimaSS(x - xreg %*% coef[narma + (1L:ncxreg)], mod)        else arimaSS(x, mod)        sigma2 <- val[[1L]][1L]/n.used    }    value <- 2 * n.used * res$value + n.used + n.used * log(2 *         pi)    aic <- if (method != "CSS")         value + 2 * sum(mask) + 2    else NA    nm <- NULL    if (arma[1L] > 0L)         nm <- c(nm, paste0("ar", 1L:arma[1L]))    if (arma[2L] > 0L)         nm <- c(nm, paste0("ma", 1L:arma[2L]))    if (arma[3L] > 0L)         nm <- c(nm, paste0("sar", 1L:arma[3L]))    if (arma[4L] > 0L)         nm <- c(nm, paste0("sma", 1L:arma[4L]))    if (ncxreg > 0L) {        nm <- c(nm, cn)        if (!orig.xreg) {            ind <- narma + 1L:ncxreg            coef[ind] <- S$v %*% coef[ind]            A <- diag(narma + ncxreg)            A[ind, ind] <- S$v            A <- A[mask, mask]            var <- A %*% var %*% t(A)        }    }    names(coef) <- nm    if (!no.optim)         dimnames(var) <- list(nm[mask], nm[mask])    resid <- val[[2L]]    tsp(resid) <- xtsp    class(resid) <- "ts"    structure(list(coef = coef, sigma2 = sigma2, var.coef = var,         mask = mask, loglik = -0.5 * value, aic = aic, arma = arma,         residuals = resid, call = match.call(), series = series,         code = res$convergence, n.cond = ncond, nobs = n.used,         model = mod), class = "Arima")})(structure(c(2510.5, 2330.7, 2050.6, 1783.3, 1632.9, 1513.1, 1525.6, 1653.1, 1475, 1567.8, 1908.5, 2587.5, 2677, 2309.5, 2246.6, 1807.2, 1522.4, 1444.4, 1598.1, 1669.2, 1494.1, 1649.1, 1701, 2120.2, 2487.6, 2242.4, 2258.4, 1881, 1611.5, 1591.4, 1748.4, 1725.7, 1542.2, 1645.9, 1913.6, 2378.9, 2700.5, 2500.3, 2197.9, 1743.5, 1514.7, 1368.4, 1600.5, 1651.6, 1428.6, 1553.2, 1753.6, 2263.7, 2675.8, 2511.1, 2100.9, 1745.2, 1573, 1483.7, 1584.9, 1578, 1482.2, 1557.2, 1782.8, 2327.7, 2561.9, 2243, 2205.8, 1724.9, 1522.6, 1534.1, 1686.6, 1695.1, 1422.5, 1428.2, 1663.4, 2326.4, 2165.3, 2144.4, 2126.4, 1681, 1526.3, 1550.9, 1758.7, 1751.7, 1462.1, 1644.2, 1765.4, 2122.8, 2475.6, 2567, 2128.8, 1810.1, 1559.1, 1555.2, 1659.9, 1896.1, 1590.5, 1627.8, 1834.5, 2399.2, 2734, 2503.4, 2278.2, 1823.9, 1576.4, 1604.2, 1708.6, 1682.9, 1460.9, 1635.8, 1868.9, 2399.7, 2729.7, 2332.5, 2170.7, 1741.3, 1504, 1527.8, 1658, 1736.5, 1575, 1666.5, 1776.2, 2491.9, 2809.8, 2481, 2142.9, 1691.8, 1617.3, 1649.5, 1825.8, 1878.9, 1637.5, 1664.9, 1973.3, 2714.1, 2888.6, 2452.4, 2230.5, 1825, 1667.4, 1657.3, 1890.5, 1891.8, 1655.6, 1744.5, 2031.9, 2541.9, 2756.2, 2500.7, 2127.8, 1953.1, 1873.8, 1868.4, 2069.8, 2008.8, 1807.2, 1901.1, 2167.8, 2503.9, 2878.8, 2567.2, 2521.1, 1967.5, 1752.5, 1742.9, 1926.3, 1927.4, 1767, 1866.8, 2316.9, 2920.8, 3204.1, 2741.2, 2557.9, 1961.7, 1810.2, 1745.4, 1881, 1933.1, 1809.3, 1912.8, 2357.5, 2679.2, 3115, 2925.2, 2591.3, 2007.9, 1858.1, 1899.9, 2067.7, 2052.7, 1901.3, 1987.3, 2249.1, 2588.2, 3091.7, 2652.3, 2356.3, 2083.8, 1965.8, 2000.7, 2186.6, 2208.4, 1947.8, 1925.2, 2159.4, 2866.3, 2900.9, 2329.1, 2529.3, 1927.9, 1896.9, 1912.1, 2132.7, 2104.9, 1922.6, 2036.9, 2351.3, 3065.8), .Tsp = c(2000, 2017.91666666667, 12), class = "ts"), order = c(2, 1, 1), seasonal = list(    order = c(1, 0, 1))): non-stationary seasonal AR part from CSS

# View the model performance on the backtesting partitions
md$leaderboard

#> Error in eval(expr, envir, enclos): object 'md' not found