Polor Plot for Time Series Object

Polor plot for time series object (ts, zoo, xts), currently support only monthly and quarterly frequency

ts_polar(ts.obj, title = NULL, width = 600, height = 600,
  left = 25, right = 25, top = 25, bottom = 25)

Arguments

ts.obj

A univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)

title

Add a title for the plot, default set to NULL

width

The widht of the plot in pixels, default set to 600

height

The height of the plot pixels, default set to 600

left

Set the left margin of the plot in pixels, default set to 25

right

Set the right margin of the plot in pixels, default set to 25

top

Set the top margin of the plot in pixels, default set to 25

bottom

Set the bottom margin of the plot in pixels, default set to 25

Examples

data(USgas)
ts_polar(USgas)