3D Surface Plot for Time Series

3D surface plot for time series object by it periodicity (currently support only monthly and quarterly frequency)

ts_surface(ts.obj)

Arguments

ts.obj

a univariate time series object of a class "ts", "zoo" or "xts" (support only series with either monthly or quarterly frequency)

Examples

ts_surface(USgas)