University of Michigan Consumer Survey, Index of Consumer Sentiment

University of Michigan Consumer Survey, Index of Consumer Sentiment: 1980 - 2019. Units: Index 1966:Q1=100

Michigan_CS

Format

Time series data - 'xts' object

Source

University of Michigan, University of Michigan: Consumer Sentiment

Examples

ts_plot(Michigan_CS)

ts_heatmap(Michigan_CS)